decomp: компонент Channels в отдельном файле
